+++ core/src/main/java/com/tbg/wms/core/model/package-info.java
@@ -13,6 +13,6 @@
  *   <li>{@link com.tbg.wms.core.model.SnapshotService} - typed snapshot serialization for debug/replay workflows.</li>
  * </ul>
  *
- * @since 1.3.1
+ * @since 1.3.2
  */
 package com.tbg.wms.core.model;