+++ core/src/main/java/com/tbg/wms/core/print/package-info.java
@@ -11,6 +11,6 @@
  *   <li>{@link com.tbg.wms.core.print.NetworkPrintService} - TCP 9100 transport for raw ZPL payloads.</li>
  * </ul>
  *
- * @since 1.3.1
+ * @since 1.3.2
  */
 package com.tbg.wms.core.print;