+++ db/src/main/java/com/tbg/wms/db/package-info.java
@@ -11,6 +11,6 @@
  *   <li>{@link com.tbg.wms.db.DbConnectivityDiagnostics} - connection diagnostics and health checks.</li>
  * </ul>
  *
- * @since 1.3.1
+ * @since 1.3.2
  */
 package com.tbg.wms.db;